oui oui le bon commit avant le partiel

+++ ERP_LaTaime/ERP_LaTaime-ejb/src/main/java/fr/miage/toulouse/m1/JEE/entities/Utilisateur.java
@@ -37,7 +37,7 @@ public class Utilisateur implements Serializable {
     
     private long numCompteBancaire;
 
-    private Long solde;
+    private Double solde;
     @OneToMany
     private List<Commande> commandes;
 
@@ -93,11 +93,11 @@ public class Utilisateur implements Serializable {
         }
     }
     
-    public Long getSolde() {
+    public Double getSolde() {
         return solde;
     }
 
-    public void setSolde(Long solde) {
+    public void setSolde(Double solde) {
         this.solde = solde;
     }
